requirements.txt
psycopg2-binary
fastapi
uvicorn
sqlalchemy
python-dotenv
asgiref==3.8.1
Django==4.1.7
django-ckeditor==6.7.1
django-js-asset==2.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
django-parler==2.3
django-jazzmin
django-rosetta==0.9.9
asyncpg
sqlalchemy-utils
Pillow